+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Create link for creating Gitlab personal access token with required scopes from Gitlab API url.
+- More descriptive message when branch does not exist on remote.
 
 ## [1.8.0] - 2023-04-23