+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## [Unreleased]
 ### Added
 - Create link for creating Gitlab personal access token with required scopes from Gitlab API url.
-- More descriptive message when branch does not exist on remote.
+- More descriptive message when git branch does not exist on remote.
 - Remove square brackets from error notification.
+- Fallback branch setting when git branch does not exist on remote.
 
 ## [1.8.0] - 2023-04-23